+++ apps/watch/src/components/Videos/Videos.spec.tsx
@@ -104,52 +104,4 @@ describe('Videos', () => {
       })
     })
   })
-  describe('list', () => {
-    it('should render a list', () => {
-      const { getByTestId } = render(
-        <MockedProvider>
-          <Videos
-            filter={{ availableVariantLanguageIds: ['529'] }}
-            layout="list"
-          />
-        </MockedProvider>
-      )
-      expect(getByTestId('videos-list')).toBeInTheDocument()
-    })
-    it('should display videos', async () => {
-      const { getByText } = render(
-        <MockedProvider
-          mocks={[
-            {
-              request: {
-                query: GET_VIDEOS,
-                variables: {
-                  where: {
-                    availableVariantLanguageIds: ['529']
-                  },
-                  offset: 0,
-                  limit: 8,
-                  languageId: '529'
-                }
-              },
-              result: {
-                data: {
-                  videos: videos
-                }
-              }
-            }
-          ]}
-        >
-          <Videos
-            filter={{ availableVariantLanguageIds: ['529'] }}
-            layout="list"
-          />
-        </MockedProvider>
-      )
-      await waitFor(() => {
-        expect(getByText(videos[0].title[0].value)).toBeInTheDocument()
-        expect(getByText(videos[7].title[0].value)).toBeInTheDocument()
-      })
-    })
-  })
 })
